Moved PrettyMixNodeBondWithLocation to models.rs
